requirements.txt
pyyaml>=6.0
click>=8.1.0
jinja2>=3.1.0
rich>=13.0.0
checkov>=3.0.0

